requirements.txt
asgiref==3.7.2
Django==4.2.4
django-jet-reboot==1.3.3
django-translations==1.3.0
Pillow==10.0.0
psycopg2-binary==2.9.7
python-dotenv==1.0.0
six==1.16.0
sqlparse==0.4.4
